#216 fix: 컴파운드 패턴 적용
