+++ src/common/component/Tab/TabList.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable import/no-named-as-default-member */
 import React, { ComponentPropsWithoutRef, ReactElement, ReactNode, cloneElement } from 'react';
 
-import { tabListStyle } from '@/common/component/Tabs/style';
+import { tabListStyle } from '@/common/component/Tab/style';
 
 interface TabListProps extends ComponentPropsWithoutRef<'ul'> {
   children: ReactNode;
@@ -9,9 +9,11 @@ interface TabListProps extends ComponentPropsWithoutRef<'ul'> {
 
 const TabList = ({ children, ...props }: TabListProps) => {
   return (
-    <ul role="tablist" css={tabListStyle} {...props}>
-      {React.Children.toArray(children).map((child, index) => cloneElement(child as ReactElement, { tabId: index }))}
-    </ul>
+    <nav>
+      <ul role="tablist" css={tabListStyle} {...props}>
+        {React.Children.toArray(children).map((child, index) => cloneElement(child as ReactElement, { tabId: index }))}
+      </ul>
+    </nav>
   );
 };
