+++ src/story/common/Tabs.stories.tsx
@@ -2,20 +2,17 @@ import type { Meta, StoryObj } from '@storybook/react';
 
 import { useState } from 'react';
 
-import Tab from '@/common/component/Tabs/Tab';
-import TabList from '@/common/component/Tabs/TabList';
-import TabPanel from '@/common/component/Tabs/TabPanel';
-import Tabs from '@/common/component/Tabs/Tabs';
+import { Tab } from '@/common/component/Tab';
 
 const meta = {
   title: 'Common/Tabs',
-  component: Tabs,
+  component: Tab,
   parameters: {
     layout: 'centered',
   },
   tags: ['autodocs'],
   args: { children: <></> },
-} satisfies Meta<typeof Tabs>;
+} satisfies Meta<typeof Tab>;
 type Story = StoryObj<typeof meta>;
 
 export default meta;
@@ -28,20 +25,20 @@ export const TabsTest: Story = {
       setSelectedTab(tabId);
     };
     return (
-      <Tabs>
-        <TabList>
-          <Tab tabId={0} selectedTab={selectedTab} onTabClick={handleTabClick} variant="round">
+      <Tab>
+        <Tab.List>
+          <Tab.Button selectedTab={selectedTab} onTabClick={handleTabClick} variant="round">
             선택된 블록
-          </Tab>
-          <Tab tabId={1} selectedTab={selectedTab} onTabClick={handleTabClick} variant="round">
+          </Tab.Button>
+          <Tab.Button selectedTab={selectedTab} onTabClick={handleTabClick} variant="round">
             전체 문서
-          </Tab>
-        </TabList>
-        <TabPanel selectedTab={selectedTab}>
+          </Tab.Button>
+        </Tab.List>
+        <Tab.Panel selectedTab={selectedTab}>
           <div>1번 탭</div>
           <div>2번 탭</div>
-        </TabPanel>
-      </Tabs>
+        </Tab.Panel>
+      </Tab>
     );
   },
 };
